refactor: move security identity resolution from interceptor to store
Simplify ConversationInterceptor by moving security identity resolution
and streaming adapter logic into ConversationStore. This makes the store
more self-contained and reduces the interceptor's responsibilities.

- Add resolveIdentity() method to ConversationStore
- Add appendAgentMessage(String, Multi<String>) overload for streaming
- Add appendAgentMessage(String, String) overload that resolves token internally
- Remove unused imports and fields from ConversationInterceptor
- Reorganize TODO.md into logical sections
